Moved MST and euclidean dist for MST to a separate algorithm file, improved naming a bit
